TIP-842: Extract attribute translation model to Akeneo\Pim\Structure
